Implement Ping on Unix.
There are two "modes" in which a Ping can be sent on Unix. If the user has
privileges to create and use raw sockets, then the "good" path is taken.
The good path uses raw sockets to send ICMP echo requests to the given
host, and then listens for ICMP echo replies. If the user does not have
privileges to do so, then the implementation simply launches a new process
using the ping or ping6 utility program (depending on whether the
destination address is an IPv4 or IPv6 address), parses the output, and
returns the result. The information from the "bad path" is less reliable,
and doesn't include things like the IPStatus property, which should be
able to cover most of the possible ICMP response message types, not just
"Success". Additionally, the "bad path" can not send arbitrary buffers
along with the echo request, although we can at least preserve the size of
the data. This means you can't inspect the PingReply and see that the same
data you sent in the echo request was returned unchanged.